Address review: move renderer to `mill.javalib.internal`, number the references
Two changes from review feedback.

Move `DepsTreeRenderer` into the `mill.javalib.internal` package and mark it
`@internal`, following `mill.javalib.internal.PublishModule`. `private[mill]`
alone still leaves the class public in the emitted bytecode, so the package
makes the intent unambiguous.

Replace the plain `(*)` elision marker with a numbered `(*n)` reference, and tag
the expanded occurrence with the same number so the two can be matched up. This
needs a pre-pass to learn which nodes are actually reached more than once, since
only those should carry a number; numbers are then assigned in printed order.

Nodes keep being identified as coursier identifies them, by the `Dependency`
they were reached with, so the same coordinates can carry more than one
reference number when they are pulled in with different exclusions. That is not
noise: `DependencyTree.Node#children` is computed from that `Dependency`, and on
the Spring/Vaadin resolution 104 of 506 nodes share a rendered label with some
other node while 26 of those genuinely have a different subtree. Keying on the
rendered label instead would point the reader at an expansion that is not
theirs. Verified that `(dependency, excluded, endorsed)` - the exact set of
fields `children` depends on - yields the same 506 classes, so nothing is lost
by keeping coursier's identity.
